chore: add xcode project + nix-ed
