--- clients/swift/cac/cac-bridging-header.h
@@ -1,6 +1,0 @@
-#ifndef CAC_Bridging_Header_h
-#define CAC_Bridging_Header_h
-
-#include "../../../headers/libcac_client.h"
-
-#endif /* CAC_Bridging_Header_h */
