--- clients/swift/exp/exp-bridging-header.h
@@ -1,6 +1,0 @@
-#ifndef Experimentation_Bridging_Header_h
-#define Experimentation_Bridging_Header_h
-
-#include "../../../headers/libexperimentation_client.h"
-
-#endif /* Experimentation_Bridging_Header_h */
