+++ clients/swift/swift/main.swift
@@ -1,0 +1,12 @@
+//
+//  main.swift
+//  swift
+//
+//  Created by naman agarwal on 27/07/24.
+//
+
+import Foundation
+
+// unit tests
+test_cac()
+test_exp()
