+++ clients/swift/swift/types.swift
@@ -1,0 +1,9 @@
+//
+//  types.swift
+//  swift
+//
+//  Created by naman agarwal on 27/07/24.
+//
+
+typealias UnknownClientPointer = OpaquePointer
+typealias Value = [String: Any]
